requirements.txt
asgiref==3.7.2
astroid==3.2.2
attrs==23.1.0
autopep8==2.0.4
beautifulsoup4==4.12.2
black==23.10.1
Brotli==1.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorclass==2.2.2
coverage==7.3.2
cryptography==41.0.5
cssbeautifier==1.14.9
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.7
Django==4.2.7
django-admin-env-notice==1.0
django-autoslug==1.9.9
django-bootstrap-icons==0.8.6
django-bootstrap-v5==1.0.11
django-cleanup==8.0.0
django-debug-toolbar==4.3.0
django-htmx==1.17.0
django-resized==1.0.2
djlint==1.34.0
docopt==0.6.2
EditorConfig==0.12.3
exceptiongroup==1.1.3
factory-boy==3.3.0
Faker==25.2.0
filelock==3.12.2
flake8==6.1.0
greenlet==3.0.3
gunicorn==21.2.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jsbeautifier==1.14.9
json5==0.9.14
Markdown==3.5.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
pep8==1.7.1
Pillow==10.1.0
pip-upgrader==1.4.15
platformdirs==3.11.0
pluggy==1.3.0
pycairo==1.25.1
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
pylint==3.2.2
pyparsing==3.1.1
pyperclip==1.8.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-xlib==0.33
pytz==2022.7.1
pyudev==0.24.1
PyYAML==6.0.1
regex==2023.3.23
requests==2.31.0
shtab==1.6.4
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.27
sqlparse==0.4.4
svgwrite==1.4.3
termcolor==2.3.0
terminaltables==3.1.10
tinycss2==1.2.1
tldr==3.2.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.2
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.7
virtualenv==20.24.6
webencodings==0.5.1
whitenoise==6.6.0
xcffib==1.5.0
xmltodict==0.13.0
